telegram_bot/requirements.txt
python-telegram-bot[ext]==21.4
httpx
python-dotenv==1.0.1 